+++ homework_17/page_objects/main_page.py
@@ -13,23 +13,23 @@ class MainPage(BasePage):
     __my_cabinet = (By.CSS_SELECTOR, "[class='head_elem singin_v']")
     __search_input = (By.XPATH, "//*[@id='search_submit']/*[@id='autocomplete']")
     __search_button = (By.CSS_SELECTOR, "#search_submit > input.search_btn")
-    __all_board_games_button = (By.XPATH, "//*[@title='Всі настільні ігри']")
+    __all_board_games_button = (By.XPATH, "//*[@href='https://desktopgames.com.ua/ua/catalog/boardgames/']")
     __main_page_items = (By.CLASS_NAME, 'product')
     __main_banner = (By.XPATH, "//*[@class='main-banners']")
     __top_menu = (By.XPATH, "//*[@class='head_cont_desk']")
 
-    def is_my_cabinet_button_displayed(self):
-        return self._wait_until_element_located(self.__my_cabinet)
+    def is_my_cabinet_button_visible(self):
+        return self._wait_until_element_visible(self.__my_cabinet)
 
-    def is_all_games_button_displayed(self):
-        return self._wait_until_element_located(self.__all_board_games_button)
+    def is_all_games_button_visible(self):
+        return self._wait_until_element_visible(self.__all_board_games_button)
 
     def click_my_cabinet_button(self):
         self._click_via_js(self.__my_cabinet)
         return LoginPage(self.driver)
 
-    def is_search_input_displayed(self):
-        return self._wait_until_element_located(self.__search_input), self._wait_until_element_located(self.__search_button)
+    def is_search_input_visible(self):
+        return self._wait_until_element_visible(self.__search_input), self._wait_until_element_visible(self.__search_button)
 
     def set_search_value(self, search_word):
         self._send_keys(locator=self.__search_input, value=search_word)
@@ -44,10 +44,10 @@ class MainPage(BasePage):
         return CatalogPage(self.driver)
 
     def are_main_page_items_displayed(self):
-        return self._wait_until_elements_are_located(self.__main_page_items)
+        return self._wait_until_elements_are_visible(self.__main_page_items)
 
-    def is_banner_displayed(self):
-        return self._wait_until_elements_are_located(self.__main_banner)
+    def is_banner_visible(self):
+        return self._wait_until_element_visible(self.__main_banner)
 
     def is_top_menu_text_displayed(self):
         return self._get_text(self.__top_menu)
